requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
click>=8.0.0
flask>=2.3.0
flask-cors>=4.0.0
colorama>=0.4.6
rich>=13.0.0
pyyaml>=6.0
jsonschema>=4.0.0
aiohttp>=3.8.0
asyncio-throttle>=1.0.0 